+++ gituser/src/commonMain/kotlin/leegroup/module/gituser/domain/repositories/GitUserRepository.kt
@@ -3,6 +3,6 @@ package leegroup.module.gituser.domain.repositories
 import leegroup.module.gituser.domain.models.GitUserModel
 
 interface GitUserRepository {
-    suspend fun getRemote(since: Int, perPage: Int): List<GitUserModel>
-    suspend fun getLocal(since: Int, perPage: Int): List<GitUserModel>
+    suspend fun getRemote(since: Long, perPage: Int): List<GitUserModel>
+    suspend fun getLocal(since: Long, perPage: Int): List<GitUserModel>
 }
